Specifies Solr v6.6.1 in SolrWrapper config
The latest version of solr_wrapper defaults to Solr v7.x, however Hyrax ships
with a default Solr schema that is not compatible with Solr 7. Eventually we
want to update the Solr schema to be compat with Solr 7, but this is easier for
now. Note that it only works becaue our Solr instance *only* comes from
solr_wrapper gem at this point.